upload group details is now parallel, update of group members doesn't load the calendar from scratch.
