+++ app/src/main/java/com/example/meetapp/GroupsDisplayActivity.java
@@ -3,6 +3,7 @@ package com.example.meetapp;
 import android.app.Dialog;
 import android.content.Intent;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -12,7 +13,11 @@ import android.widget.Button;
 import android.widget.Toast;
 
 import com.firebase.ui.firestore.FirestoreRecyclerOptions;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.CollectionReference;
+import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.Query;
 
@@ -30,6 +35,7 @@ public class GroupsDisplayActivity extends AppCompatActivity {
     private String userID;
     GroupsDisplayFeaturesHandler groupsDisplayFeaturesHandler;
     private FirebaseFirestore db = FirebaseFirestore.getInstance();
+    private DocumentReference usersRef;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -38,8 +44,8 @@ public class GroupsDisplayActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
         setToolbar();
         setDialogs();
-        setCreateNewGroupListener();
         setRecyclerViewAndAdapter();
+        readUserNameFromDB();
     }
 
     private void fetchUserIdAndUserName(){
@@ -69,6 +75,20 @@ public class GroupsDisplayActivity extends AppCompatActivity {
     protected void onStop() {
         super.onStop();
         adapter.stopListening();
+    }
+
+    private void readUserNameFromDB() {
+        usersRef = db.collection("users").document(userID);
+        usersRef.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+            @Override
+            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
+                if (task.isSuccessful()) {
+                    DocumentSnapshot document = task.getResult();
+                    userName = (String) document.get("name");
+                    setCreateNewGroupListener();
+                }
+            }
+        });
     }
 
     private void setCreateNewGroupListener(){
